Stop asking repeatdly to turn bluetooth on
Whenever BT is OFF, the ScanFragment would request to turn on and while
it was not enabled the app is unusable.

The bug was caused because when onResume() was called it requested the
user to turn the bluetooth on, which caused the activity to call
onPause(). If the user refused to turn bluetooth on, onResume() was
called once again which requested the user to turn on the bluetooth and
so on.