Fix bug that caused simultaneous bluetooth scans
Fix bug that caused bluetooth to have multiple scans simultaneosly. The
bug was caused when one scan fragment started scanning before the other
scan fragment called stopScan, which caused the current fragment callback
to be discarted instead. The solution was to give each fragment a
scanCallback object and send this object to the bluetoothWrapper to
start/stop scanning from the fragments personal callback.