Fix logic in feedbackOfGatewayListState
Fix feedbackOfGatewatListState logic to turn visibility on when the
service list is empty and off when the service list isn't empty.

The current version turned the visibility on when the list wasn't empty
which led to a visual feedback of "Looking for gateways..." when there
were already some items in the gateway list.