Select tab farthest to the right
The fragment opens the tab that is farthest to the right as default.

The UI is positioned in a way that the most used tabs are in the far
right. Since these tabs are the one that the user interacts the most, it
makes sense that these are the tabs that the fragment opens at.