Fix white background at login tab
Gives each fragment the responsibility to select the appropriate
background color, text color and selectTabIndicator color.

Since these fragments share a common xml, using the standard xml
control component of tabs is confusing and error prone. A better
approach would be to use different xml's for each activity, but this
would take a lot of time and effort and wouldn't aggregate any immediate
value for the product.